Remove deprecated DmfsTask class and move tests to new location
