+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/tasks/DmfsTaskList.kt
@@ -17,7 +17,7 @@ import at.bitfire.synctools.storage.toContentValues
 import org.dmfs.tasks.contract.TaskContract
 import java.util.LinkedList
 import java.util.logging.Logger
-import at.bitfire.ical4android.DmfsTask as LegacyDmfsTask
+
 
 /**
  * Represents a locally stored task list, containing [DmfsTask]s (tasks).
@@ -137,17 +137,6 @@ class DmfsTaskList(
             throw LocalStorageException("Couldn't query task entity", e)
         }
         return null
-    }
-
-    /**
-     * Gets a specific task, identified by its ID, from this task list.
-     *
-     * @param id    task ID
-     * @return task (or `null` if not found)
-     */
-    fun getLegacyTask(id: Long): LegacyDmfsTask? {
-        val entity = getTaskEntity(id, null) ?: return null
-        return LegacyDmfsTask(this, entity.entityValues)
     }
 
     /**
